refactor(ImagePreviewer): abstract ImagePreviewer component
This component is aimed to display the image and listen to mouse events
to draw rects.
